+++ README.md
@@ -1,6 +1,6 @@
-[![Build Status](https://travis-ci.org/caststack/python-mpd-parser.svg?branch=master)](https://travis-ci.org/caststack/python-mpd-parser)
+[![Build Status](https://travis-ci.org/caststack/python-mpegdash.svg?branch=master)](https://travis-ci.org/caststack/python-mpegdash)
 
-# python-mpd-parser
+# python-mpegdash
 MPEG-DASH MPD(Media Presentation Description) Parser
 compatible with Python2.6+ and Python3
 
@@ -9,16 +9,16 @@ compatible with Python2.6+ and Python3
     $ python3 -m unittest discover
 
 ## Usage
-    from mpd.parser import MPDParser
-    
+    from mpegdash.parser import MPEGDASHParser
+
     # Parse from file path
     mpd_path = './tests/mpd-samples/sample-001.mpd'
-    mpd = MPDParser.parse(mpd_path)
-    
+    mpd = MPEGDASHParser.parse(mpd_path)
+
     # Parse from url
     mpd_url = 'http://yt-dash-mse-test.commondatastorage.googleapis.com/media/motion-20120802-manifest.mpd'
-    mpd = MPDParser.parse(mpd_url)
-    
+    mpd = MPEGDASHParser.parse(mpd_url)
+
     # Parse from string
     mpd_string = '''
     <MPD xmlns="urn:mpeg:DASH:schema:MPD:2011" mediaPresentationDuration="PT0H1M52.43S" minBufferTime="PT1.5S"
@@ -36,8 +36,7 @@ compatible with Python2.6+ and Python3
       </Period>
     </MPD>
     '''
-    mpd = MPDParser.parse(mpd_string)
-    
+    mpd = MPEGDASHParser.parse(mpd_string)
+
     # Write to xml file
-    MPDParser.write(mpd, './tests/mpd-samples/output.mpd')
-    
+    MPEGDASHParser.write(mpd, './tests/mpd-samples/output.mpd')
