+++ mpegdash/parser.py
@@ -1,4 +1,3 @@
-from __future__ import unicode_literals
 from xml.dom import minidom
 
 # python3 support
@@ -7,11 +6,11 @@ try:
 except:
     from urllib.request import urlopen
 
-from mpd.nodes import MPD
-from mpd.utils import parse_child_nodes, write_child_node
+from mpegdash.nodes import MPEGDASH
+from mpegdash.utils import parse_child_nodes, write_child_node
 
 
-class MPDParser(object):
+class MPEGDASHParser(object):
     @classmethod
     def load_xmldom(cls, string_or_url):
         if '<MPD' in string_or_url:
@@ -28,7 +27,7 @@ class MPDParser(object):
     @classmethod
     def parse(cls, string_or_url):
         xml_root_node = cls.load_xmldom(string_or_url)
-        return parse_child_nodes(xml_root_node, 'MPD', MPD)[0]
+        return parse_child_nodes(xml_root_node, 'MPD', MPEGDASH)[0]
 
     @classmethod
     def write(cls, mpd, filepath):
